Added new sections and a toggle for colour pallette
Added in a section for showing my supplementary materials for any papers I'll work on. I might even retroactively go back and make some for papers I've been involved in. Could be fun.

and the coolest part, sorted out a toggle for colour scheme. Not the impressive, but I got a kick out of it